+++ Project 5/SDLProject 5/SDLProject/Map.cpp
@@ -1,7 +1,7 @@
 #include "Map.h"
 
 
-Map::Map(int width, int height, unsigned int *levelData, GLuint textureID, float tile_size, int
+Map::Map(int width, int height, float *levelData, GLuint textureID, float tile_size, int
 tile_count_x, int tile_count_y)
 {
     this->width = width;
